Factor out test code out of glutil package.
The test code had some cgo imports that were not in a _test.go file
(because cgo cannot be used in _test.go files), which means any package
importing glutil would incur those imports (even if unused). Factor out
the test into ./tests package and unexport all other unsupported
functionality.
