+++ tests/context_darwin_amd64.go
@@ -4,7 +4,7 @@
 
 // +build darwin
 
-package glutil
+package tests
 
 // TODO(crawshaw): Only used in glutil tests for now (cgo is not support in _test.go files).
 // TODO(crawshaw): Export some kind of Context. Work out what we can offer, where. Maybe just for tests.
